move CommonTlsContext to its own file
