+++ src/core/xds/grpc/xds_common_types.h
@@ -34,52 +34,6 @@
 
 namespace grpc_core {
 
-struct CommonTlsContext {
-  struct CertificateProviderPluginInstance {
-    std::string instance_name;
-    std::string certificate_name;
-
-    bool operator==(const CertificateProviderPluginInstance& other) const {
-      return instance_name == other.instance_name &&
-             certificate_name == other.certificate_name;
-    }
-
-    std::string ToString() const;
-    bool Empty() const;
-  };
-
-  struct CertificateValidationContext {
-    struct SystemRootCerts {
-      bool operator==(const SystemRootCerts&) const { return true; }
-    };
-    std::variant<std::monostate, CertificateProviderPluginInstance,
-                 SystemRootCerts>
-        ca_certs;
-    std::vector<StringMatcher> match_subject_alt_names;
-
-    bool operator==(const CertificateValidationContext& other) const {
-      return ca_certs == other.ca_certs &&
-             match_subject_alt_names == other.match_subject_alt_names;
-    }
-
-    std::string ToString() const;
-    bool Empty() const;
-  };
-
-  CertificateValidationContext certificate_validation_context;
-  CertificateProviderPluginInstance tls_certificate_provider_instance;
-
-  bool operator==(const CommonTlsContext& other) const {
-    return certificate_validation_context ==
-               other.certificate_validation_context &&
-           tls_certificate_provider_instance ==
-               other.tls_certificate_provider_instance;
-  }
-
-  std::string ToString() const;
-  bool Empty() const;
-};
-
 struct XdsExtension {
   // The type, either from the top level or from inside the TypedStruct.
   absl::string_view type;
